se agregan los formularios de creación de proyecto y evento tambien  se agregan las lineas de investigación

+++ app/Http/Controllers/Semilleros.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Models\Semillero;
+use Illuminate\Support\Facades\Storage;
 
 class Semilleros extends Controller
 {
@@ -22,7 +23,7 @@ class Semilleros extends Controller
         $semillero->fecCreaSemillero = $r->input('fecha');
         $semillero->presSemillero = $presPath;
         $semillero->resCreaSemillero = $resPath;
-        $semillero->lineaSemillero = $r->input('linea_investigacion');
+        $semillero->lineaSemillero = $r->input('selLinea');
         $semillero->descSemillero = $r->input('descripcion');
         $semillero->misSemillero = $r->input('mision');
         $semillero->visSemillero = $r->input('vision');
@@ -35,24 +36,30 @@ class Semilleros extends Controller
     public function actualizar(Request $r, $id){
         $semillero = Semillero::findOrFail($id);
         if ($r->hasFile('logo')) {
-            Storage::delete($semillero->logoSemillero);    
+            if (Storage::exists($semillero->logoSemillero)) {
+                Storage::delete($semillero->logoSemillero);
+            }
             $logoPath = $r->file('logo')->store('Semilleros/logos', 'public');
         }else {
             $logoPath = $semillero->logoSemillero;
         }
         if ($r->hasFile('presentacion')) {
-            Storage::delete($semillero->presSemillero);    
+            if (Storage::exists($semillero->presSemillero)) {
+                Storage::delete($semillero->presSemillero);    
+            }
             $presPath = $r->file('presentacion')->store('Semilleros/presentaciones', 'public');
         }else {
             $presPath = $semillero->presSemillero;
         }
         if ($r->hasFile('resolucion')) {
-            Storage::delete($semillero->resCreaSemillero);
+            if (Storage::exists($semillero->resCreaSemillero)) {
+                Storage::delete($semillero->resCreaSemillero);
+            }
             $resPath = $r->file('resolucion')->store('Semilleros/resoluciones', 'public');
         }else {
             $resPath = $semillero->resCreaSemillero;
         }
-        $semillero->nomSemillero = $id;
+        $semillero->nomSemillero = $r->input('nombreHidden');
         $semillero->emailSemillero = $r->input('correo');
         $semillero->logoSemillero = $logoPath;
         $semillero->fecCreaSemillero = $r->input('fecha');
